move URL validation methods from UIUtils to AppUtil class to support batch mode
UIUtils depends on SWT. Moving URL validation code to AppUtil supports batch mode execution without requiring SWT in the CLASSPATH.
